requirements.txt
pandas
numpy
scikit-learn
lightgbm
requests
streamlit
python-dotenv
